src/requirements.txt
argh==0.26.2
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
pathtools==0.1.2
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
python-Levenshtein==0.12.0
pytz==2020.4
PyYAML==5.3.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
watchdog==0.10.4
wrapt==1.12.1
